TT-999: Switch to async / await syntax, improve type safety
Also moved the API types into their own folder so we don't have to have
circular dependencies on the things that use them.

Authors: @richardtowers
