+++ lib/create-response-handler.ts
@@ -11,7 +11,7 @@
  */
 /** */
 
-import { TranslatedResponseBody, Scenario } from './passport-verify-strategy'
+import { TranslatedResponseBody, Scenario } from './verify-service-provider-api/translated-response-body'
 
 /**
  * ResponseScenarios
