+++ lib/verify-service-provider-api/authn-request-response.ts
@@ -1,0 +1,9 @@
+/**
+ * Represents the JSON returned by a success response from the
+ * `/generate-request` endpoint in the verify-service-provider
+ */
+export interface AuthnRequestResponse {
+  samlRequest: string,
+  requestId: string,
+  ssoLocation: string
+}
