+++ lib/verify-service-provider-api/error-message.ts
@@ -1,0 +1,7 @@
+/**
+ * Represents the JSON returned by all error responses from the verify-service-provider
+ */
+export interface ErrorMessage {
+  code: number,
+  message: string
+}
